oauth-svc/requirements.txt
appdirs==1.4.3
bottle==0.12.13
packaging==16.8
pyparsing==2.2.0
six==1.10.0
pycrypto==2.6.1
python-jwt==2.0.1

